test: move wallet utils and add createAddress option.
